simplified typeatom structure
