+++ grakn-graql/src/main/java/ai/grakn/graql/internal/pattern/property/HasResourceTypeProperty.java
@@ -29,7 +29,6 @@ import ai.grakn.graql.admin.VarAdmin;
 import ai.grakn.graql.VarName;
 import ai.grakn.graql.internal.gremlin.EquivalentFragmentSet;
 import ai.grakn.graql.internal.query.InsertQueryExecutor;
-import ai.grakn.graql.internal.reasoner.atom.binary.SelectableTypeAtom;
 import ai.grakn.graql.internal.reasoner.atom.binary.TypeAtom;
 import ai.grakn.util.ErrorMessage;
 import ai.grakn.util.Schema;
