+++ grakn-graql/src/main/java/ai/grakn/graql/internal/pattern/property/PlaysRoleProperty.java
@@ -30,7 +30,6 @@ import ai.grakn.graql.VarName;
 import ai.grakn.graql.internal.gremlin.EquivalentFragmentSet;
 import ai.grakn.graql.internal.gremlin.fragment.Fragments;
 import ai.grakn.graql.internal.query.InsertQueryExecutor;
-import ai.grakn.graql.internal.reasoner.atom.binary.SelectableTypeAtom;
 import ai.grakn.graql.internal.reasoner.atom.binary.TypeAtom;
 import ai.grakn.graql.internal.reasoner.atom.predicate.IdPredicate;
 import com.google.common.collect.Sets;
@@ -123,6 +122,6 @@ public class PlaysRoleProperty extends AbstractVarProperty implements NamedPrope
         IdPredicate predicate = getIdPredicate(typeVariable, typeVar, vars, parent);
 
         VarAdmin resVar = Graql.var(varName).playsRole(Graql.var(typeVariable)).admin();
-        return new SelectableTypeAtom(resVar, predicate, parent);
+        return new TypeAtom(resVar, predicate, parent);
     }
 }
