+++ grakn-graql/src/main/java/ai/grakn/graql/internal/pattern/property/SubProperty.java
@@ -28,7 +28,6 @@ import ai.grakn.graql.VarName;
 import ai.grakn.graql.internal.gremlin.EquivalentFragmentSet;
 import ai.grakn.graql.internal.gremlin.fragment.Fragments;
 import ai.grakn.graql.internal.query.InsertQueryExecutor;
-import ai.grakn.graql.internal.reasoner.atom.binary.SelectableTypeAtom;
 import ai.grakn.graql.internal.reasoner.atom.binary.TypeAtom;
 import ai.grakn.graql.internal.reasoner.atom.predicate.IdPredicate;
 import ai.grakn.util.ErrorMessage;
@@ -125,6 +124,6 @@ public class SubProperty extends AbstractVarProperty implements NamedProperty, U
         IdPredicate predicate = getIdPredicate(typeVariable, typeVar, vars, parent);
 
         VarAdmin resVar = Graql.var(varName).sub(Graql.var(typeVariable)).admin();
-        return new SelectableTypeAtom(resVar, predicate, parent);
+        return new TypeAtom(resVar, predicate, parent);
     }
 }
